append files
