+++ README copy.md
@@ -1,0 +1,36 @@
+ "To-do-list" is a tool that helps to organize your day. It simply lists the things that you need to do and allows you to mark them as complete.
+ 
+ ## Delopyment
+ Click [here](https://alfred-kinyua.github.io/to-do-list-project-/) to check out the online version
+
+ ## Built With
+- Languages - HTML, CSS and Javascript
+- Frameworks - None
+- Technologies - Web Technology
+- Other tools - webpack
+- npm as software packages system
+# Getting started
+### Prerequisites
+One should have knowledge about the following
+- HTML, CSS, Javascript and npm
+
+### Run
+- Navigate to the local folder and clone the repository from the github 
+- Run "git clone [url](git@github.com:Alfred-KInyua/to-do-list-project-.git)"
+- Navigate to the cloned project
+- On your terminal Run "npm start"
+
+# Contributing
+Feel free to check out the [Issues page](https://github.com/Alfred-KInyua/to-do-list-project-/issues)
+
+# Show your support
+Give a star if you like this project
+
+# Acknowledgement
+ [Microverse](https://www.microverse.org/?grsf=i6yi2m)
+ 
+# License
+This project is MIT licensed.
+
+# Author
+1. [Alfred Kinyua](https://github.com/Alfred-KInyua)
