+++ src/index.html
@@ -1,0 +1,31 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="UTF-8">
+  <meta http-equiv="X-UA-Compatible" content="IE=edge">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <script src="https://kit.fontawesome.com/ebe56293cc.js" crossorigin="anonymous"></script>
+
+  <title>My To-do-list program</title>
+</head>
+<body>
+  <div class="wrapper">
+    <div class="wrap">
+      <h2>Today's To Do</h2>
+      <i class="fa-solid fa-arrows-rotate reset"></i>
+    </div>
+
+    <form class="form">
+      <input type="text" required placeholder="Add to your list" id="run">
+      <button type="submit" id="check"><i class="fa fa-check"></i></button>
+    </form>
+
+    <ul class="list"></ul>
+
+    <div class="footer-section">
+      <a href="" class="clear">Clear all completed</a>
+    </div>
+
+  </div>
+</body>
+</html>
