+++ src/modules/application.js
@@ -1,88 +1,3 @@
-..................................................................
-index.html
-..................................................................
-
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta http-equiv="X-UA-Compatible" content="IE=edge">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <script src="https://kit.fontawesome.com/ebe56293cc.js" crossorigin="anonymous"></script>
-
-  <title>My To-do-list program</title>
-</head>
-<body>
-  <div class="wrapper">
-    <div class="wrap">
-      <h2>Today's To Do</h2>
-      <i class="fa-solid fa-arrows-rotate reset"></i>
-    </div>
-
-    <form class="form">
-      <input type="text" required placeholder="Add to your list" id="run">
-      <button type="submit" id="check"><i class="fa fa-check"></i></button>
-    </form>
-
-    <ul class="list"></ul>
-
-    <div class="footer-section">
-      <a href="" class="clear">Clear all completed</a>
-    </div>
-
-  </div>
-</body>
-</html>
-
-
-
-................................
-javascript 
-................................
-
-
-
-
-import MyTodoClass from './modules/application.js';
-import './style.css';
-
-const input = document.querySelector('#run');
-const button = document.querySelector('#check');
-
-MyTodoClass.display();
-
-input.addEventListener('focus', (e) => {
-  e.preventDefault();
-  button.style.display = 'block';
-});
-
-document.querySelector('#check').addEventListener('click', (e) => {
-  e.preventDefault();
-  if (input.value) {
-    const desc = input.value;
-    const done = new MyTodoClass(desc);
-    done.addtodolist();
-  }
-  return 0;
-});
-
-document.querySelectorAll('.clear').forEach((link) => {
-  link.addEventListener('click', (e) => {
-    e.preventDefault();
-    MyTodoClass.clearCompleted();
-  });
-});
-
-document.querySelector('.reset').addEventListener('click', (e) => {
-  e.preventDefault();
-  MyTodoClass.reset();
-});
-
-
-
-
-
-
 export default class MyTodoClass {
   constructor(desc) {
     this.desc = desc;
@@ -234,118 +149,4 @@ export default class MyTodoClass {
       });
     });
   };
-}
-
-
-
-
-............................................
-CSS
-.............................................
-@import url('https://fonts.googleapis.com/css2?family=Poppins:wght@200&display=swap');
-
-body {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  font-family: 'Poppins', Arial, Helvetica, sans-serif;
-  background-color: gray;
-}
-
-.wrapper {
-  width: 50%;
-  background: white;
-}
-
-.line {
-  border: 0.2px solid #e8e8e9;
-}
-
-.wrap {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 0.9em 2em;
-  border-bottom: 0.2px solid #e8e8e9;
-}
-
-form {
-  display: flex;
-  justify-content: space-between;
-  padding: 24px;
-  border-bottom: 0.2px solid #e8e8e9;
-}
-
-.form-group textarea {
-  padding: 6px;
-  width: 90%;
-  display: inline-block;
-  border: none;
-  resize: none;
-  outline-width: 0.2px;
-}
-
-form button {
-  background-color: transparent;
-  background-repeat: no-repeat;
-  border: none;
-  display: none;
-  font-size: 1.5rem;
-}
-
-.textarea.line-through {
-  text-decoration: line-through;
-}
-
-form input::placeholder {
-  font-style: italic;
-}
-
-.list {
-  list-style-type: none;
-}
-
-.list-item {
-  padding: 4px 24px;
-  display: flex;
-  justify-content: space-between;
-  border-bottom: 0.2px solid #e8e8e9;
-}
-
-form input {
-  outline: none;
-  border: none;
-}
-
-.form-group {
-  flex: 2;
-  width: 100%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.form-group input {
-  margin-right: 5px;
-}
-
-#editInput {
-  display: none;
-}
-
-.action-icons i {
-  margin: 0 1em;
-}
-
-.footer-section {
-  padding: 12px;
-  background: #f3f3f3;
-}
-
-.footer-section a {
-  text-decoration: none;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  color: inherit;
 }
