+++ src/style.css
@@ -1,0 +1,107 @@
+@import url('https://fonts.googleapis.com/css2?family=Poppins:wght@200&display=swap');
+
+body {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  font-family: 'Poppins', Arial, Helvetica, sans-serif;
+  background-color: gray;
+}
+
+.wrapper {
+  width: 50%;
+  background: white;
+}
+
+.line {
+  border: 0.2px solid #e8e8e9;
+}
+
+.wrap {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 0.9em 2em;
+  border-bottom: 0.2px solid #e8e8e9;
+}
+
+form {
+  display: flex;
+  justify-content: space-between;
+  padding: 24px;
+  border-bottom: 0.2px solid #e8e8e9;
+}
+
+.form-group textarea {
+  padding: 6px;
+  width: 90%;
+  display: inline-block;
+  border: none;
+  resize: none;
+  outline-width: 0.2px;
+}
+
+form button {
+  background-color: transparent;
+  background-repeat: no-repeat;
+  border: none;
+  display: none;
+  font-size: 1.5rem;
+}
+
+.textarea.line-through {
+  text-decoration: line-through;
+}
+
+form input::placeholder {
+  font-style: italic;
+}
+
+.list {
+  list-style-type: none;
+}
+
+.list-item {
+  padding: 4px 24px;
+  display: flex;
+  justify-content: space-between;
+  border-bottom: 0.2px solid #e8e8e9;
+}
+
+form input {
+  outline: none;
+  border: none;
+}
+
+.form-group {
+  flex: 2;
+  width: 100%;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.form-group input {
+  margin-right: 5px;
+}
+
+#editInput {
+  display: none;
+}
+
+.action-icons i {
+  margin: 0 1em;
+}
+
+.footer-section {
+  padding: 12px;
+  background: #f3f3f3;
+}
+
+.footer-section a {
+  text-decoration: none;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  color: inherit;
+}
